+++ src/main/java/com/nhnacademy/bookpubshop/product/repository/ProductRepositoryCustom.java
@@ -1,5 +1,6 @@
 package com.nhnacademy.bookpubshop.product.repository;
 
+import com.nhnacademy.bookpubshop.product.dto.GetProductListForOrderResponseDto;
 import com.nhnacademy.bookpubshop.product.dto.response.GetProductByTypeResponseDto;
 import com.nhnacademy.bookpubshop.product.dto.response.GetProductDetailResponseDto;
 import com.nhnacademy.bookpubshop.product.dto.response.GetProductListResponseDto;
